Diary 型 fix
