+++ src/components/DiaryViewer/index.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import styled from "styled-components";
 
 import BreakPoint from "../../constants/BreakPoint";
-import { Diary } from "../../store/diaries/reducers";
+import { Diary } from "../../server/services/diaries/types";
 import Button, { buttonTheme } from "../Button";
 import DiaryController from "../DiaryController";
