+++ src/server/services/diaries/ShowDiaryService.ts
@@ -2,16 +2,16 @@ import { Request, Response } from "@specter/client";
 import { Service } from "@specter/specter";
 
 import { getDiaryFromFirestore } from "../../../lib/firestore";
-import { Diary } from "../../../store/diaries/reducers";
+import { Diary } from "./types";
 
-type RequestHeader = Record<string, any>;
-type RequestBody = Record<string, any>;
+type RequestHeader = Record<string, string | string[]>;
+type RequestBody = Record<string, unknown>;
 export type ShowDiaryServiceQuery = {
   firestore: FirebaseFirestore.Firestore;
   userId: string;
   diaryId: string;
 };
-type ResponseHeader = Record<string, any>;
+type ResponseHeader = Record<string, unknown>;
 export type ShowDiaryServiceBody = Diary;
 
 export default class ShowDiaryService extends Service {
@@ -25,14 +25,6 @@ export default class ShowDiaryService extends Service {
     try {
       // TODO loggerを入れる
       const diary = await getDiaryFromFirestore(request.query);
-      if (!diary) {
-        // TODO エラーレスポンス生成関数を作る
-        // TODO firebaseのエラー型を確認する
-        const response = new Response<ResponseHeader, any>({}, diary);
-        response.setStatus(404);
-        return response;
-      }
-
       const response = new Response<ResponseHeader, ShowDiaryServiceBody>(
         {},
         diary
