+++ src/server/services/diaries/types.ts
@@ -1,0 +1,7 @@
+export type Diary = {
+  id: string;
+  title: string;
+  body: string;
+  imageUrls: string[];
+  lastEdited: string;
+};
